RIA-7911 Updating hearing requirements to adjustments - notifications
