--- src/functionalTest/resources/scenarios/RIA-7689-internal-update-hearing-requirements-non-ada-notification.json
@@ -1,68 +1,0 @@
-{
-  "description": "RIA-7689 Internal update hearing requirements notification - non-ADA",
-  "request": {
-    "uri": "/asylum/ccdAboutToSubmit",
-    "credentials": "Judge",
-    "input": {
-      "id": 7689,
-      "eventId": "updateHearingRequirements",
-      "state": "preHearing",
-      "caseData": {
-        "template": "minimal-internal-appeal-submitted.json",
-        "replacements": {
-          "detentionFacility": "immigrationRemovalCentre",
-          "ircName": "Brookhouse",
-          "appellantInDetention": "Yes",
-          "isAcceleratedDetainedAppeal": "No",
-          "notificationAttachmentDocuments": [
-            {
-              "id": "1",
-              "value": {
-                "tag": "internalHearingRequirementsUpdatedLetter",
-                "document": {
-                  "document_url": "{$FIXTURE_DOC1_PDF_URL}",
-                  "document_binary_url": "{$FIXTURE_DOC1_PDF_URL_BINARY}",
-                  "document_filename": "{$FIXTURE_DOC1_PDF_FILENAME}"
-                },
-                "suppliedBy": "",
-                "description": "",
-                "dateUploaded": "{$TODAY}"
-              }
-            }
-          ]
-        }
-      }
-    }
-  },
-  "expectation": {
-    "status": 200,
-    "errors": [],
-    "caseData": {
-      "template": "minimal-internal-appeal-submitted.json",
-      "replacements": {
-        "detentionFacility": "immigrationRemovalCentre",
-        "ircName": "Brookhouse",
-        "appellantInDetention": "Yes",
-        "isAcceleratedDetainedAppeal": "No",
-        "notificationsSent": [
-          {
-            "id": "7689_INTERNAL_HEARING_REQUIREMENTS_CHANGED_DET",
-            "value": "$/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/"
-          }
-        ]
-      }
-    },
-    "notifications": [
-      {
-        "reference": "7689_INTERNAL_HEARING_REQUIREMENTS_CHANGED_DET",
-        "recipient": "det-irc-brookhouse@example.com",
-        "subject": "IAFT – SERVE IN PERSON: Hearing requirements changed",
-        "body": [
-          "Home Office reference: A1234567",
-          "HMCTS reference: PA/12345/2019",
-          "Appellant name: Talha Awan"
-        ]
-      }
-    ]
-  }
-}
