+++ src/functionalTest/resources/scenarios/RIA-7911-internal-update-hearing-adjustments-non-ada-notification.json
@@ -1,0 +1,91 @@
+{
+  "description": "RIA-7911 Internal update hearing adjustments notification - non-ADA",
+  "request": {
+    "uri": "/asylum/ccdAboutToSubmit",
+    "credentials": "Judge",
+    "input": {
+      "id": 7911,
+      "eventId": "updateHearingAdjustments",
+      "state": "preHearing",
+      "caseData": {
+        "template": "minimal-internal-appeal-submitted.json",
+        "replacements": {
+          "listCaseHearingCentre": "bradford",
+          "listCaseHearingLength": "60",
+          "listCaseHearingDate": "2018-12-31T12:34:56",
+          "ariaListingReference": "AA/12345/1234",
+          "detentionFacility": "immigrationRemovalCentre",
+          "ircName": "Brookhouse",
+          "appellantInDetention": "Yes",
+          "isAcceleratedDetainedAppeal": "No",
+          "notificationAttachmentDocuments": [
+            {
+              "id": "1",
+              "value": {
+                "tag": "internalHearingAdjustmentsUpdatedLetter",
+                "document": {
+                  "document_url": "{$FIXTURE_DOC1_PDF_URL}",
+                  "document_binary_url": "{$FIXTURE_DOC1_PDF_URL_BINARY}",
+                  "document_filename": "{$FIXTURE_DOC1_PDF_FILENAME}"
+                },
+                "suppliedBy": "",
+                "description": "",
+                "dateUploaded": "{$TODAY}"
+              }
+            }
+          ]
+        }
+      }
+    }
+  },
+  "expectation": {
+    "status": 200,
+    "errors": [],
+    "caseData": {
+      "template": "minimal-internal-appeal-submitted.json",
+      "replacements": {
+        "listCaseHearingCentre": "bradford",
+        "listCaseHearingLength": "60",
+        "listCaseHearingDate": "2018-12-31T12:34:56",
+        "ariaListingReference": "AA/12345/1234",
+        "detentionFacility": "immigrationRemovalCentre",
+        "ircName": "Brookhouse",
+        "appellantInDetention": "Yes",
+        "isAcceleratedDetainedAppeal": "No",
+        "notificationsSent": [
+          {
+            "id": "7911_CHANGE_TO_HEARING_REQUIREMENTS_ADMIN_OFFICER",
+            "value": "$/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/"
+          },
+          {
+            "id": "7911_INTERNAL_HEARING_ADJUSTMENTS_CHANGED_DET",
+            "value": "$/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/"
+          }
+        ]
+      }
+    },
+    "notifications": [
+      {
+        "reference": "7911_CHANGE_TO_HEARING_REQUIREMENTS_ADMIN_OFFICER",
+        "recipient": "ao-review@example.com",
+        "subject": "Accelerated detained appeal: change to hearing requirements",
+        "body": [
+          "HMCTS reference: PA/12345/2019",
+          "Listing reference: AA/12345/1234",
+          "Appellant name: Talha Awan",
+          "The online service: https://manage-case.platform.hmcts.net/"
+        ]
+      },
+      {
+        "reference": "7911_INTERNAL_HEARING_ADJUSTMENTS_CHANGED_DET",
+        "recipient": "det-irc-brookhouse@example.com",
+        "subject": "ADA – SERVE IN PERSON: Hearing requirements changed",
+        "body": [
+          "HMCTS reference: PA/12345/2019",
+          "Home Office reference: A1234567",
+          "Appellant name: Talha Awan"
+        ]
+      }
+    ]
+  }
+}
