+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/entities/ccd/Event.java
@@ -109,7 +109,6 @@ public enum Event {
     MARK_APPEAL_AS_DETAINED("markAppealAsDetained"),
     MARK_AS_READY_FOR_UT_TRANSFER("markAsReadyForUtTransfer"),
     UPDATE_DETENTION_LOCATION("updateDetentionLocation"),
-    UPDATE_HEARING_REQUIREMENTS("updateHearingRequirements"),
 
     @JsonEnumDefaultValue
     UNKNOWN("unknown");
