--- src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/detentionengagementteam/DetentionEngagementTeamUpdateHearingRequirementsPersonalisation.java
@@ -1,84 +1,0 @@
-package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.detentionengagementteam;
-
-import static java.util.Objects.requireNonNull;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.DocumentTag.INTERNAL_HEARING_REQUIREMENTS_UPDATED_LETTER;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.utils.AsylumCaseUtils.getLetterForNotification;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.utils.AsylumCaseUtils.isAcceleratedDetainedAppeal;
-
-import com.google.common.collect.ImmutableMap;
-import java.io.IOException;
-import java.util.*;
-import lombok.extern.slf4j.Slf4j;
-import org.json.JSONObject;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.*;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.EmailWithLinkNotificationPersonalisation;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.service.DetEmailService;
-import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.DocumentDownloadClient;
-import uk.gov.service.notify.NotificationClientException;
-
-@Slf4j
-@Service
-public class DetentionEngagementTeamUpdateHearingRequirementsPersonalisation implements EmailWithLinkNotificationPersonalisation {
-
-    private final String internalUpdateHearingRequirementsTemplateId;
-    private final DocumentDownloadClient documentDownloadClient;
-    private final DetEmailService detEmailService;
-    private String adaPrefix;
-    private String nonAdaPrefix;
-    private final String ircValue = "immigrationRemovalCentre";
-
-    public DetentionEngagementTeamUpdateHearingRequirementsPersonalisation(
-        @Value("${govnotify.template.updateHearingRequirements.detentionEngagementTeam.email}") String internalUpdateHearingRequirementsTemplateId,
-        DetEmailService detEmailService,
-        DocumentDownloadClient documentDownloadClient,
-        @Value("${govnotify.emailPrefix.adaInPerson}") String adaPrefix,
-        @Value("${govnotify.emailPrefix.nonAdaInPerson}") String nonAdaPrefix
-    ) {
-        this.internalUpdateHearingRequirementsTemplateId = internalUpdateHearingRequirementsTemplateId;
-        this.detEmailService = detEmailService;
-        this.documentDownloadClient = documentDownloadClient;
-        this.adaPrefix = adaPrefix;
-        this.nonAdaPrefix = nonAdaPrefix;
-    }
-
-    @Override
-    public String getTemplateId() {
-        return internalUpdateHearingRequirementsTemplateId;
-    }
-
-    @Override
-    public Set<String> getRecipientsList(AsylumCase asylumCase) {
-        return detEmailService.getRecipientsList(asylumCase);
-    }
-
-    @Override
-    public String getReferenceId(Long caseId) {
-        return caseId + "_INTERNAL_HEARING_REQUIREMENTS_CHANGED_DET";
-    }
-
-    @Override
-    public Map<String, Object> getPersonalisationForLink(AsylumCase asylumCase) {
-        requireNonNull(asylumCase, "asylumCase must not be null");
-
-        return ImmutableMap
-            .<String, Object>builder()
-            .put("subjectPrefix", isAcceleratedDetainedAppeal(asylumCase) ? adaPrefix : nonAdaPrefix)
-            .put("appealReferenceNumber", asylumCase.read(AsylumCaseDefinition.APPEAL_REFERENCE_NUMBER, String.class).orElse(""))
-            .put("homeOfficeReferenceNumber", asylumCase.read(AsylumCaseDefinition.HOME_OFFICE_REFERENCE_NUMBER, String.class).orElse(""))
-            .put("appellantGivenNames", asylumCase.read(AsylumCaseDefinition.APPELLANT_GIVEN_NAMES, String.class).orElse(""))
-            .put("appellantFamilyName", asylumCase.read(AsylumCaseDefinition.APPELLANT_FAMILY_NAME, String.class).orElse(""))
-            .put("documentLink", getInternalUpdateHearingRequirementsDocumentInJsonObject(asylumCase))
-            .build();
-    }
-
-    private JSONObject getInternalUpdateHearingRequirementsDocumentInJsonObject(AsylumCase asylumCase) {
-        try {
-            return documentDownloadClient.getJsonObjectFromDocument(getLetterForNotification(asylumCase, INTERNAL_HEARING_REQUIREMENTS_UPDATED_LETTER));
-        } catch (IOException | NotificationClientException e) {
-            log.error("Failed to get Internal Hearing Requirements changed document in compatible format", e);
-            throw new IllegalStateException("Failed to get Internal Hearing Requirements changed document in compatible format");
-        }
-    }
-}
